style!: Move to conventional search & replace api
This code is much more readable, though I regret that it's not as fast.

Documentation was also updated.

BREAKING CHANGE: page `"sway:crafting"` `filter` was removed in favor of naming
the HBox `"sway_crafting_hbox"`.
BREAKING CHANGE: `Form` now renames the element that wraps the children to
`"content"`. This element is a VBox.